requirements.txt
flask
gunicorn
flask-mail
mysql.connector
psycopg2
requests